+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 2.0.3 - 2022-10-10
+* change return type of tokenGetter to Future
+
+## 2.0.2 - 2022-10-10
+* update the document
+
+## 2.0.1 - 2022-09-27
+* update the document
+
+## 2.0.0 - 2022-09-27
+* add batch upload object
+* add batch delete object
+
 ## 1.0.5 - 2022-08-06
 * add README_ZH.md